requirements.txt
pytest==3.7.0
pygithub==1.40
